+++ HISTORY.rst
@@ -3,6 +3,18 @@ History
 -------
 
 
+13.3.0 (2024-04-27)
+++++++++++++++++++
+
+- Update WakaTime.Shared.ExtensionUtils dependency to v4.2.2.
+
+
+13.2.0 (2023-10-26)
+++++++++++++++++++
+
+- Update WakaTime.Shared.ExtensionUtils dependency to v4.2.1.
+
+
 13.0.2 (2022-12-18)
 ++++++++++++++++++